Develop

See merge request DTS/phenolic!6

+++ phenolic_api/src/main/java/pt/ipb/phenolic/repos/MoleculeRepository.java
@@ -3,5 +3,5 @@ package pt.ipb.phenolic.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.ipb.phenolic.models.Molecule;
 
-public interface MoleculeRepository extends JpaRepository<Molecule, Integer> {
+public interface MoleculeRepository extends JpaRepository<Molecule, Long> {
 }
